add refresh token endpoint

+++ src/utils/cloudinary.js
@@ -18,11 +18,15 @@ const uploadOnCloudinary =async (localFilePath) => {
         resource_type: "auto"
     })
     // file has been uploaded
-    console.log("file is uploaded on cloudinary", response.url);
+    //console.log("file is uploaded on cloudinary", response.url);
+    fs.unlinkSync(localFilePath)
     return response;
     
    } catch (error) {
+    console.log("cloudinary error:", error);
+    if(localFilePath){
     fs.unlinkSync(localFilePath)// remove the locally saved temporiry file as the upload operation got failed
+    }
     return null;
    }    
 }
